Работа друзьями в пользователях и лайками в фильмах
Исправлены все замечания

Took 1 hour 41 minutes

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -4,83 +4,91 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.controller.exception.*;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.FilmService;
-import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 @Slf4j
 @RestController
 @RequestMapping("/films")
 public class FilmController {
-    @Autowired
-    FilmStorage filmStorage;
+    private static final LocalDate MIN_DATE_RELEASE = LocalDate.of(1895, 12, 28);
     @Autowired
     FilmService filmService;
 
     @PostMapping()
-    public ResponseEntity<Film> addFilm(@RequestBody Film film) {
-        filmStorage.addFilm(film);
+    public Film addFilm(@RequestBody Film film) throws ValidationException {
+        generalValidateFilm(film);
+        filmService.addFilm(film);
         log.warn("Фильм добавлен: " + film);
-        return new ResponseEntity<>(film, HttpStatus.OK);
+        new ResponseEntity<>(film, HttpStatus.OK);
+        return film;
     }
 
     @PutMapping()
-    public ResponseEntity<Film> updateFilm(@RequestBody Film film) {
-        filmStorage.updateFilm(film);
+    public Film updateFilm(@RequestBody Film film) throws ValidationException {
+        generalValidateFilm(film);
+        filmService.updateFilm(film);
         log.warn("Фильм изменён: " + film);
-        return new ResponseEntity<>(film, HttpStatus.OK);
+        new ResponseEntity<>(film, HttpStatus.OK);
+        return film;
     }
 
     @GetMapping()
     public List<Film> getAllFilms() {
-        return new ArrayList<>(filmStorage.getAllFilms());
+        return new ArrayList<>(filmService.getAllFilms());
     }
 
     @GetMapping(value = "/{id}")
-    public ResponseEntity<?> getFilm(@PathVariable int id) {
-        Film film = filmStorage.getFilmById(id);
-        return new ResponseEntity<>(film, HttpStatus.OK);
+    public Film getFilm(@PathVariable int id) {
+        Film film = filmService.getFilmById(id);
+        new ResponseEntity<>(film, HttpStatus.OK);
+        return film;
     }
 
     @PutMapping(value = "/{id}/like/{userId}")
-    public ResponseEntity<?> addLike(@PathVariable int id, @PathVariable int userId) {
+    public void addLike(@PathVariable int id, @PathVariable int userId) {
         filmService.addLike(id, userId);
-        return new ResponseEntity<>(HttpStatus.OK);
+        new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping(value = "/{id}/like/{userId}")
-    public ResponseEntity<List<User>> deleteLike(@PathVariable int id, @PathVariable int userId) {
+    public void deleteLike(@PathVariable int id, @PathVariable int userId) {
+        if (userId < 1) {
+            throw new NotFoundException("Ошибка userId < 1 :" + userId);
+        }
         filmService.deleteLike(id, userId);
-        return new ResponseEntity<>(HttpStatus.OK);
+        new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping(value = "/popular")
-    public ResponseEntity<List<Film>> getPopularFilms(@RequestParam(required = false) Integer count) {
-        if (count == null) {
-            count = 10;
+    public List<Film> getPopularFilms(@RequestParam(required = false, defaultValue = "10") Integer count) {
+        List<Film> filmList = new ArrayList<>(filmService.getPopularFilms(count));
+        new ResponseEntity<>(filmList, HttpStatus.OK);
+        return filmList;
+    }
+
+    private void generalValidateFilm(Film film) throws ValidationException {
+
+        if (StringUtils.isEmpty(film.getName()))
+            throw new ValidationException("Название не может быть пустым!");
+
+        if (film.getDescription().length() > 200)
+            throw new ValidationException("Максимальная длина описания — 200 символов!");
+
+        if (film.getReleaseDate().isBefore(FilmController.MIN_DATE_RELEASE)) {
+            throw new ValidationException("Дата релиза — не раньше 28.12.1985 !");
         }
-        return new ResponseEntity<>(filmService.getPopularFilms(count), HttpStatus.OK);
-    }
+        if (film.getDuration() < 1) {
+            throw new ValidationException("Продолжительность фильма должна быть положительной!");
+        }
 
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    private Map<String, String> handlerValidationException(final ValidationException e) {
-        log.warn("Ошибка ввода данных: " + e.getMessage());
-        return Map.of("error", e.getMessage());
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    private Map<String, String> handlerNotFoundException(final NotFoundException e) {
-        log.warn("Ошибка ввода данных: " + e.getMessage());
-        return Map.of("error", e.getMessage());
     }
 
 }
